requirements.txt
hazelcast-python-client
flask
python-consul